refact: move applicant namespace constant to Constants class
