+++ src/main/java/uk/gov/companieshouse/SuppressionApplication.java
@@ -6,8 +6,6 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class SuppressionApplication {
 
-    public static final String APPLICATION_NAMESPACE = "suppression-api";
-
     public static void main (String[] args) {
         SpringApplication.run(SuppressionApplication.class, args);
     }
