+++ src/main/java/uk/gov/companieshouse/config/LoggerConfig.java
@@ -5,7 +5,7 @@ import org.springframework.context.annotation.Configuration;
 import uk.gov.companieshouse.logging.Logger;
 import uk.gov.companieshouse.logging.LoggerFactory;
 
-import static uk.gov.companieshouse.SuppressionApplication.APPLICATION_NAMESPACE;
+import static uk.gov.companieshouse.model.Constants.APPLICATION_NAMESPACE;
 
 @Configuration
 public class LoggerConfig {
