+++ src/main/java/uk/gov/companieshouse/model/Constants.java
@@ -2,12 +2,11 @@ package uk.gov.companieshouse.model;
 
 public class Constants {
 
-   public static final String APPLICATION_RECEIVED_MESSAGE_TYPE = "suppression_application_received";
-   public static final String APPLICATION_RECEIVED_EMAIL_SUBJECT = "Application received: %s";
+    public static final String APPLICATION_NAMESPACE = "suppression-api";
 
-   public static final String APPLICATION_CONFIRMATION_MESSAGE_TYPE = "suppression_application_confirmation";
-   public static final String APPLICATION_CONFIRMATION_EMAIL_SUBJECT = "Application to remove your home address from the Companies House register submitted: %s";
+    public static final String APPLICATION_RECEIVED_MESSAGE_TYPE = "suppression_application_received";
+    public static final String APPLICATION_RECEIVED_EMAIL_SUBJECT = "Application received: %s";
 
-
-    
+    public static final String APPLICATION_CONFIRMATION_MESSAGE_TYPE = "suppression_application_confirmation";
+    public static final String APPLICATION_CONFIRMATION_EMAIL_SUBJECT = "Application to remove your home address from the Companies House register submitted: %s";
 }
