+++ src/test/java/uk/gov/companieshouse/controller/PaymentControllerTest_PATCH.java
@@ -10,7 +10,6 @@ import org.springframework.test.web.servlet.MockMvc;
 import uk.gov.companieshouse.email_producer.EmailSendingException;
 import uk.gov.companieshouse.fixtures.SuppressionFixtures;
 import uk.gov.companieshouse.logging.Logger;
-import uk.gov.companieshouse.logging.Logger;
 import uk.gov.companieshouse.model.Suppression;
 import uk.gov.companieshouse.model.payment.PaymentPatchRequest;
 import uk.gov.companieshouse.model.payment.PaymentStatus;
